backend/requirements.txt
# Core FastAPI dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Vector search and embeddings
sentence-transformers==2.2.2
faiss-cpu==1.7.4
torch==2.2.2
transformers==4.36.0

# BM25 text search
rank-bm25==0.2.2

# Async and utilities
asyncio-throttle==1.0.2
python-multipart==0.0.6
httpx==0.25.2

# Data processing
numpy==1.24.3
pandas==2.1.4

# Logging and monitoring
structlog==23.2.0

# Development dependencies
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
isort==5.12.0
mypy==1.7.1

# Optional: Database drivers (for future use)
# sqlalchemy==2.0.23
# asyncpg==0.29.0
# redis==5.0.1